python/contrib/drivers/ustreamer/pyproject.toml
[project]
name = "jumpstarter-driver-ustreamer"
dynamic = ["version", "urls"]
description = ""
authors = [
  { name = "Nick Cao", email = "ncao@redhat.com" },
  { name = "Miguel Angel Ajo Pelayo", email = "majopela@redhat.com" },
  { name = "Kirk Brauer", email = "kbrauer@hatci.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jumpstarter",
    "pillow>=10.4.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]

[tool.uv.sources]
jumpstarter = { workspace = true }

[tool.hatch.metadata.hooks.vcs.urls]
Homepage = "https://jumpstarter.dev"
source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"

[tool.hatch.version]
source = "vcs"
raw-options = { 'root' = '../../../'}

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
